+++ cmd/autobump/dig.go
@@ -24,15 +24,15 @@ func injectAppContext() *internal.AppInternal {
 	return appInternal
 }
 
-func injectSingleController() *controllers.SingleController {
+func injectLocalController() *controllers.LocalController {
 	container := dig.New()
 
 	if err := internal.RegisterProviders(container); err != nil {
 		panic(err)
 	}
 
-	var controller *controllers.SingleController
-	if err := container.Invoke(func(c *controllers.SingleController) {
+	var controller *controllers.LocalController
+	if err := container.Invoke(func(c *controllers.LocalController) {
 		controller = c
 	}); err != nil {
 		panic(err)