+++ CHANGELOG.md
@@ -16,6 +16,10 @@ Exceptions are acceptable depending on the circumstances (critical bug fixes tha
 
 ## [Unreleased]
 
+### Changed
+
+- changed the Go version to `1.26.1` and updated all module dependencies
+
 ## [2.20.0] - 2026-03-09
 
 ### Added